Move core.core to tools.core
